[SP] Use size_t types instead of int and unsigned int when comparing to sizeof operator or STL template size function.

+++ code/cgame/cg_credits.cpp
@@ -575,7 +575,7 @@ qboolean CG_Credits_Draw( void )
 			//
 			iYpos += iFontHeight*2;	// skip blank line then move to main pos
 			//
-			for (unsigned int i=0; i<CreditCard.vstrText.size(); i++)
+			for (size_t i=0; i<CreditCard.vstrText.size(); i++)
 			{
 				StringAndSize_t &StringAndSize = CreditCard.vstrText[i];
 				iWidth = StringAndSize.GetPixelLength();
@@ -648,7 +648,7 @@ qboolean CG_Credits_Draw( void )
 
 						// now print any dotted members...
 						//
-						for (unsigned int i=0; i<CreditLine.vstrText.size(); i++)
+						for (size_t i=0; i<CreditLine.vstrText.size(); i++)
 						{
 							StringAndSize_t &StringAndSize = CreditLine.vstrText[i];
 							iWidth = StringAndSize.GetPixelLength();
